Convert useStoredCards to TypeScript and move to own file
